Readability fix

+++ test/utils/ProtocolHarness.sol
@@ -9,7 +9,7 @@ import { Protocol } from "../../src/Protocol.sol";
 contract ProtocolHarness is Protocol {
     constructor(address spogRegistrar_, address mToken_) Protocol(spogRegistrar_, mToken_) {}
 
-    function mintNonce() external view returns (uint48) {
+    function mintNonce() external view returns (uint256) {
         return _mintNonce;
     }
